requirements.txt
alembic==1.18.4
fastapi==0.135.1
fastapi-cache2==0.2.2
fastapi-cli==0.0.24
fastapi-cloud-cli==0.14.1
fastapi-users==15.0.4
fastapi-users-db-sqlalchemy==7.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
SQLAlchemy==2.0.48
uvicorn==0.41.0
asyncpg==0.31.0
sqlmodel==0.0.37
hashids==1.3.1
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.1.0
pytest-cov==7.1.0
pytest-asyncio==1.3.0
locust
